+++ tests/cli/versions.rs
@@ -11,23 +11,37 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
+//! Tests of the `conserve versions` command.
+
 use assert_cmd::prelude::*;
+use predicates::function::function;
 
 use crate::run_conserve;
 
 #[test]
 fn versions() {
     run_conserve()
-        .args(&["versions", "testdata/archive/simple/v0.6.10"])
+        .args(&["versions", "--utc", "testdata/archive/simple/v0.6.10"])
         .assert()
         .success()
         .stdout(
             "\
-b0000                complete   2021-03-04 05:21:15     0:00
-b0001                complete   2021-03-04 05:21:30     0:00
-b0002                complete   2021-03-04 05:27:28     0:00
+b0000                complete   2021-03-04 13:21:15     0:00
+b0001                complete   2021-03-04 13:21:30     0:00
+b0002                complete   2021-03-04 13:27:28     0:00
 ",
         );
+}
+
+#[test]
+fn versions_in_local_time() {
+    // Without --utc we don't know exactly what times will be produced,
+    // and it's hard to control the timezone for tests on Windows.
+    run_conserve()
+        .args(&["versions", "testdata/archive/simple/v0.6.10"])
+        .assert()
+        .success()
+        .stdout(function(|s: &str| s.lines().count() == 3));
 }
 
 #[test]
@@ -48,14 +62,19 @@ b0002
 #[test]
 fn versions_sizes() {
     run_conserve()
-        .args(&["versions", "--sizes", "testdata/archive/simple/v0.6.10"])
+        .args(&[
+            "versions",
+            "--sizes",
+            "--utc",
+            "testdata/archive/simple/v0.6.10",
+        ])
         .assert()
         .success()
         .stdout(
             "\
-b0000                complete   2021-03-04 05:21:15     0:00           0 MB
-b0001                complete   2021-03-04 05:21:30     0:00           0 MB
-b0002                complete   2021-03-04 05:27:28     0:00           0 MB
+b0000                complete   2021-03-04 13:21:15     0:00           0 MB
+b0001                complete   2021-03-04 13:21:30     0:00           0 MB
+b0002                complete   2021-03-04 13:27:28     0:00           0 MB
 ",
         );
 }